main: split user_buttons.c
